feat: move `CustomRetrievalEvaluator` in `evaluation` module
