+++ colpali_engine/evaluation/__init__.py
@@ -1,0 +1,1 @@
+from .retrieval_evaluator import CustomRetrievalEvaluator
