+++ colpali_engine/evaluation/retrieval_evaluator.py
