+++ scripts/infer/run_inference_with_python.py
@@ -7,8 +7,8 @@ from torch.utils.data import DataLoader
 from tqdm import tqdm
 from transformers import AutoProcessor
 
+from colpali_engine.evaluation.retrieval_evaluator import CustomRetrievalEvaluator
 from colpali_engine.models import ColPali
-from colpali_engine.trainer.retrieval_evaluator import CustomRetrievalEvaluator
 from colpali_engine.utils.processing_utils import BaseVisualRetrieverProcessor
